Added markdownize Liquid filter
I wanted a Markdown equivalent to the textilize filter, so I can make
an excerpt in a post's YAML header and process it with something like
{{ page.excerpt | markdownize }}. This initial implementation is
woefully slapdash; it pretty much hardcodes RDiscount as its Markdown
processor, because that's the one I prefer.

I looked for an easy way to follow the site's configuration and select
RDiscount or Maruku depending on user settings, but Jekyll::Filters
doesn't have any obvious way to get at the Jekyll::Site object that
holds all the settings. Jekyll::Filters is passed along as a raw module
to Liquid, which makes the Filters module independent (good) but
disconnected (bad).

I'll look for a way to dig out the site's Markdown configuration some
time when I'm not digesting a large turkey dinner.