Implemented <!--more--> divider, passes test
A post (and only a post; not a page or a layout) with <!--more--> in
it (case-insensitive, okay to have spaces around 'more'), is divided
into excerpt and remainder parts, exposed on the post as page.excerpt
and page.remainder, respectively. The content of the page has a div
marker (<div id="more" /></div>) inserted so links can go directly to
the remainder portion of a post.