Added fixes, improvements to WordPress converter
Fixed: Posts with an excerpt containing nothing but whitespace (for
whatever reason, a bunch on my old WP blog had nothing in them but a
single newline character) created invalid YAML headers in the imported
posts, which in turn prevented those posts from being processed by
Jekyll. Added a whitespace-detecting regular expression to the delete_if method used to remove empty fields from the YAML output.

New: Added format parameter to process method. This makes it possible to
specify 'textile' as the file extension for converted posts.

New: Added tags handling. The converter finds tags assigned to posts and
adds them to the YAML header for each post. Note that this only picks up
tags from WordPress 2.3 and later; it uses the taxonomy system in the
newest versions of WP and doesn't have a clue how to handle the earlier
database structure.

Also changed database code to be more idiomatic Ruby. The Sequel gem has
lovely Ruby syntax for forming datasets from chained method calls, so
there's no need for the old-school technique of creating a raw SQL
string and passing it to the database object.